model data in `models.ts`file
